requirements.txt
Flask
Flask-SocketIO
pytest
pytest-mock
python-dotenv
paho-mqtt
pymongo
Werkzeug
flask-jwt-extended
boto3
eventlet
gunicorn
